main.js разделен на отдельные модули
